Revert "[CONTROLLER/GRPC] moves synchronizer service from trisolaris to grpc"
This reverts commit a24d9b636156feab37ebc19ef0dadffdf703cf32.
